WIP interacting with JavaScript in Rust

+++ src/apu/audio_formatter.rs
@@ -1,5 +1,5 @@
 pub fn mix_audio(input: f32, src: f32, volume: u16) -> f32 {
-    let max_volume = 1.0 / 128.0 as f32;
+    let max_volume = 1.0 / 128.0_f32;
     let max_audio = f32::MAX as f64;
     let min_audio = f32::MIN as f64;
     let src1 = src * volume as f32 * max_volume;
