+++ src/gpu/lcd/screen.rs
@@ -28,8 +28,8 @@ impl Screen {
         let y_index = ((SCREEN_WIDTH * y) as usize) * COLOR_PER_PIXEL;
         let x_index = y_index + ((x as usize) * COLOR_PER_PIXEL); // For RGB
         self.pixels[x_index as usize] = color.0;
-        self.pixels[x_index + 1 as usize] = color.1;
-        self.pixels[x_index + 2 as usize] = color.2;
+        self.pixels[x_index + 1_usize] = color.1;
+        self.pixels[x_index + 2_usize] = color.2;
     }
 }
